before refactoring

+++ WebTest/Entities/Organization.cs
@@ -8,6 +8,6 @@ namespace WebTest.Entities
     public class Organization
     {
         public int Id { get; set; }
-        public int Inn { get; set; }
+        public long Inn { get; set; }
     }
 }
